Fix ApplyLeave API to match DotNetCode logic: - Return HTTP 200 for all responses (LOP flow fix for frontend) - Change noOfDays entity type from Integer to Double (half-day support) - Skip carry-forward balance update for LOP leaves - Fix EL Monday check to use original leaveTypeId - Match .NET validation order and error messages

+++ java/src/main/java/com/officeconnect/entity/EmpLeaveApplication.java
@@ -30,7 +30,7 @@ public class EmpLeaveApplication {
     private Date toDate;
 
     @Column(name = "Duration")
-    private Integer noOfDays;
+    private Double noOfDays;
 
     @Column(name = "Reason")
     private String reason;
@@ -110,8 +110,8 @@ public class EmpLeaveApplication {
     public Date getToDate() { return toDate; }
     public void setToDate(Date toDate) { this.toDate = toDate; }
 
-    public Integer getNoOfDays() { return noOfDays; }
-    public void setNoOfDays(Integer noOfDays) { this.noOfDays = noOfDays; }
+    public Double getNoOfDays() { return noOfDays; }
+    public void setNoOfDays(Double noOfDays) { this.noOfDays = noOfDays; }
 
     public String getReason() { return reason; }
     public void setReason(String reason) { this.reason = reason; }
